# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api transferência simplificada",
        "description": "API que simula um sistema simplificado de transferências",
        "contact": {
            "name": "D.Venancio",
            "email": "venanciomagalhaesd@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Criação de um novo usuário",
                "operationId": "439154bae968629faa2b1f4b04646e68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "document",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "document": {
                                        "description": "Valid CPF or CNPJ (numbers only)",
                                        "type": "string",
                                        "example": "41590444094"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Senha123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Senha123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "0af3db3a-3cfc-447b-95c1-1c1466e94c71"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Deividson Venancio Magalhaes"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "common"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "02215725621"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "venanciomagalhsaesd@gmail.com"
                                                },
                                                "amount": {
                                                    "type": "string",
                                                    "example": "1000.00"
                                                },
                                                "permissions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "send_transaction"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Permission to send transactions",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already exists (duplicate email or document) or Invalid document",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This user already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Field validation errors",
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email has already been taken."
                                                ],
                                                "password": [
                                                    "The password confirmation does not match."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ]
}